requirements.txt
aiogram>=3.4.1
httpx>=0.27.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
pdfplumber>=0.11.4
pandas>=2.2.2
openpyxl>=3.1.5
python-dotenv>=1.0.1
pydantic>=2.7.4
rank-bm25>=0.2.2
chromadb>=0.5.3
fastapi>=0.112.0
uvicorn>=0.30.1
orjson>=3.10.6
tabulate>=0.9.0
pytest>=8.2.1